pyproject.toml
[project]
name = "wsnetbuilder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arcgis>=2.3.0.2",
    "bmi-arcgis-restapi>=2.4.9",
    "folium>=0.18.0",
    "geopandas>=1.0.1",
    "ipykernel>=6.29.5",
    "leafmap>=0.38.16",
    "owslib>=0.32.0",
    "requests>=2.32.3",
    "setuptools>=75.3.0",
    "streamlit>=1.40.0",
]
